Store deleted state on Entity

+++ src/Entity.php
@@ -29,6 +29,11 @@ abstract class Entity {
     protected $columns;
 
     /**
+     * @var bool
+     */
+    protected $deleted = false;
+
+    /**
      * @var string
      */
     protected static $table = "";
@@ -271,6 +276,15 @@ abstract class Entity {
     }
 
     /**
+     * Whether this item has been deleted from the database.
+     *
+     * @return bool
+     */
+    public function isDeleted(): bool {
+        return $this->deleted;
+    }
+
+    /**
      * Simple factory method to create a new entity instance and optionally set the values.
      *
      * @param $data array|null
@@ -379,6 +393,8 @@ abstract class Entity {
     }
 
     /**
+     * Load Entity(ies) from the Database where a column equals/in $value.
+     *
      * @param $column string
      * @param $value string|int|array
      * @param $limit int|string|null
@@ -407,7 +423,9 @@ abstract class Entity {
     }
 
     /**
-     * @param $id int[]|int
+     * Load Entity(ies) from the Database where Id column equals/in $id.
+     *
+     * @param $id int[]|string[]|int|string
      * @return \JPI\ORM\Entity\Collection|static|null
      */
     public static function getById($id) {
@@ -422,15 +440,17 @@ abstract class Entity {
      * @return void
      */
     public function reload(): void {
-        if ($this->isLoaded()) {
-            $row = static::select("*", $this->getId());
-            if ($row) {
-                $this->setValues($row, true);
-                return;
-            }
-
-            $this->setId(null);
+        if (!$this->isLoaded() || $this->isDeleted()) {
+            return;
         }
+
+        $row = static::select("*", $this->getId());
+        if ($row) {
+            $this->setValues($row, true);
+            return;
+        }
+
+        $this->setId(null);
     }
 
     /**
@@ -471,6 +491,10 @@ abstract class Entity {
      */
     public function save(): bool {
         if ($this->isLoaded()) {
+            if ($this->isDeleted()) {
+                return false;
+            }
+
             $rowsAffected = static::newQuery()->update($this->getValuesToSave(), $this->getId());
             if ($rowsAffected === 0) {
                 // Updating failed so reset id
@@ -504,11 +528,11 @@ abstract class Entity {
      * @return bool Whether or not deletion was successful.
      */
     public function delete(): bool {
-        if ($this->isLoaded()) {
+        if ($this->isLoaded() && !$this->isDeleted()) {
             $rowsAffected = static::newQuery()->delete($this->getId());
-            return $rowsAffected > 0;
+            $this->deleted = $rowsAffected > 0;
         }
 
-        return false;
+        return $this->deleted;
     }
 }
